TEST-#2722: add ASV read_csv skiprows benchmark
